+++ database/migrations/2023_01_24_210051_add_uuid_column_to_failed_jobs_table.php
@@ -27,6 +27,18 @@ return new class () extends Migration {
      */
     public function down(): void
     {
+        if (!Schema::hasTable('failed_jobs') || !Schema::hasColumn('failed_jobs', 'uuid')) {
+            return;
+        }
+
+        try {
+            Schema::table('failed_jobs', function (Blueprint $table) {
+                $table->dropUnique('failed_jobs_uuid_unique');
+            });
+        } catch (\Throwable) {
+
+        }
+
         Schema::table('failed_jobs', function (Blueprint $table) {
             $table->dropColumn('uuid');
         });
